img-search-git/requirements.txt
fastapi
uvicorn
python-multipart
numpy
faiss-cpu
torch --index-url https://download.pytorch.org/whl/cpu
torchvision --index-url https://download.pytorch.org/whl/cpu
ftfy
regex
git+https://github.com/openai/CLIP.git
Pillow